added Burger component
